requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
django-cors-headers>=4.0.0

python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
django-filter==23.5
django-extensions==3.2.3
pillow==10.1.0
django-storages==1.14.2
boto3==1.29.7  # For S3 storage (optional)

# Email
django-ses==3.5.0  # AWS SES
sendgrid==6.11.0  # SendGrid alternative

# CRM Integration
requests==2.31.0
simple-salesforce==1.12.5  # Salesforce
hubspot-api-client>=7.5.0  # HubSpot

# Security & Validation
django-ratelimit==4.1.0
django-honeypot==0.8.0
django-recaptcha==3.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Admin Enhancements
django-admin-list-filter-dropdown==1.0.3
django-import-export==3.3.4

# Testing
factory-boy==3.3.0
coverage==7.3.2
freezegun==1.2.2  # For time-based testing

# API Documentation
drf-spectacular==0.27.0  # OpenAPI 3.0 schema generation

# A/B Testing
scipy>=1.11.0  # For statistical significance calculations
numpy>=1.24.0  # Required by scipy
